Globally convert PCM to use frame counts instead of size
Stage 1 of implementing additional sample depths. Counts are easier
to deal with than byte sizes. Adopt some Alsa terminology for
consistency.

Change any names to help avoid confusion. Fix any PCM driver problems
I find in a target I don't have that I can deal with. Either test or
I forge ahead and maybe break some eggs (which seems to be the only
efficient way to get anything done ;).

Give the PCM peak interface a slight makeover since it's being changed
anyway. Peak functions no longer use static data internally and thus
are usable by multiple callers without mutual interference. Gently
suggest the possibility of more than two channels in the data struct.

Increment minimum plugin and codec versions since this breaks
compatibility.

+++ lib/rbcodec/dsp/dsp_filter.h
@@ -49,8 +49,8 @@ void filter_hs_coefs(unsigned long cutoff, unsigned long Q, long db,
                      struct dsp_filter *f);
 void filter_copy(struct dsp_filter *dst, const struct dsp_filter *src);
 void filter_flush(struct dsp_filter *f);
-void filter_process(struct dsp_filter *f, int32_t * const buf[], int count,
-                    unsigned int channels);
+void filter_process(struct dsp_filter *f, int32_t * const buf[],
+                    unsigned long count, unsigned int channels);
 /* ring buffer */
 void enqueue(int32_t var, int32_t* buffer, int *head, int boundary);
 int32_t dequeue(int32_t* buffer, int *head, int boundary);
